Resume Scanner  — NLP-Based Semantic Matching/requirements.txt
pdfminer.six
nltk
spacy